+++ App.js
@@ -9,6 +9,8 @@ import { ApolloProvider } from "react-apollo";
 import client from "./src/modules/apollo-client";
 import AppNavigator from "./src/navigation";
 
+console.log(AppNavigator);
+
 class App extends React.Component {
   static navigatorOptions = { title: "Home" };
