+++ src/constants/routes.js
@@ -3,3 +3,4 @@ export const APP = "APP";
 export const STORYBOOK = "STORYBOOK";
 export const CREATE_ACCOUNT = "CREATE_ACCOUNT";
 export const LOGIN = "LOGIN";
+export const NOT_LOGGED_IN = "NOT_LOGGED_IN";
