+++ src/navigation/NonLoggedInStack/Home/index.js
@@ -1,4 +1,4 @@
-import { Home } from "../../screens";
+import { Home } from "../../../screens";
 
 const home = {
   HOME: {
